Wrap componentRegistry components in AppContext and break infinite loop
